+++ .mocharc.json
@@ -4,5 +4,5 @@
   "require": "ts-node/register"
 }
 /**
- * @url https://github.com/mochajs/mocha/blob/master/example/config/.mocharc.json
+ * @see https://github.com/mochajs/mocha/blob/master/example/config/.mocharc.json
  */
