+++ _pages/_app.tsx
@@ -3,7 +3,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { AppProps } from 'next/app';
 import { trpc } from '../hooks/trpc';
-import { supabase } from '../infra/supabase';
+// import { supabase } from '../infra/supabase';
 import './globals.css';
 
 const queryClient = new QueryClient({
@@ -20,8 +20,8 @@ function App({ Component, pageProps }: AppProps) {
   return (
     <QueryClientProvider client={queryClient}>
       {/* <Auth.UserContextProvider supabaseClient={supabase}> */}
-        <Component {...pageProps} />
-        <ReactQueryDevtools initialIsOpen={false} />
+      <Component {...pageProps} />
+      <ReactQueryDevtools initialIsOpen={false} />
       {/* </Auth.UserContextProvider> */}
     </QueryClientProvider>
   );
