+++ components/CustomPopover.tsx
@@ -1,6 +1,6 @@
 import { XMarkIcon } from '@heroicons/react/24/solid';
 import { Root, Trigger, Content, Close } from '@radix-ui/react-popover';
-import { IconButton } from './IconButton';
+import { IconButton } from '../app/_components/IconButton';
 
 type Placement = 'top' | 'right' | 'bottom' | 'left';
