+++ components/JsonImport.tsx
@@ -2,7 +2,7 @@ import { ArrowUpIcon, CheckIcon } from '@heroicons/react/24/solid';
 import { DayOneData, toEntry } from '../domain/DayOneEntry';
 import { Entry } from '../domain/Entry';
 import { useJsonImport } from '../hooks/useJsonImport';
-import { IconButton } from './IconButton';
+import { IconButton } from '../app/_components/IconButton';
 import { Spinner } from './Spinner';
 
 export const JsonImport = (props: {
